Move scripts to a `bin/` directory
In order to be proper compatible with the way Homebrew expects to
install binaries the scripts need to be moved in a bin directory and
dynamically search the prefix in the `ktx` script in order to use the
helper commands.

Also removed the manual installation instructions for `ktx` since that
wouldn't work due to the missing helper command.